Vuelta atras, que se rompio todo

+++ trunk/Ejercicio 6/v5 dist. varios brokers con anillo y lider/Anillo/inet.h
@@ -67,7 +67,7 @@ typedef struct{
 
 
 typedef struct{
-    uint64_t idBroker;
+    int idBroker;
     int estado;
 }MsgLider_t;
