+++ trunk/Ejercicio 6/v5 dist. varios brokers con anillo y lider/Anillo/listener.c
@@ -18,12 +18,12 @@ extern int recibir (int, void*, size_t);
 main(int argc, char *argv[])
 {
     key_t key = ftok("/tmp/buchwaldipcs", SHM_VERSION);
-    int shmversion = shmget(key, sizeof(uint64_t) , 0660 | IPC_CREAT);
-    uint64_t* id = (uint64_t*)shmat(shmversion, NULL, 0);   
+    int shmversion = shmget(key, sizeof(int) , 0660 | IPC_CREAT);
+    int* id = (int*)shmat(shmversion, NULL, 0);   
     
     char mostrar[300];
     
-    sprintf(mostrar, "[BROKER] --> ID: %lu\t| pid: %d\n", *id, getpid());
+    sprintf(mostrar, "[BROKER] --> ID: %d\t| pid: %d\n", *id, getpid());
     write(fileno(stdout), mostrar, strlen(mostrar));
     
     struct sigaction sa; 
@@ -333,7 +333,7 @@ main(int argc, char *argv[])
                write(fileno(stdout),mostrar,strlen(mostrar));
                exit(1);
            }else{
-               sprintf(mostrar,"[RECIBIDO] <-- ID: %lu\t%s\n",msgLider.idBroker, imprimirCodigo(msgLider.estado));
+               sprintf(mostrar,"[RECIBIDO] <-- ID: %d\t%s\n",msgLider.idBroker, imprimirCodigo(msgLider.estado));
                write(fileno(stdout),mostrar,strlen(mostrar));
 
                if((msgLider.idBroker > *id)&&(msgLider.estado!=FIN)){
@@ -365,12 +365,12 @@ main(int argc, char *argv[])
                         exit(1);
                    }
 			   }
-			   sprintf(mostrar,"[ENVIADO] --> id: %lu\t%s\n",msgLider.idBroker, imprimirCodigo(msgLider.estado));
+			   sprintf(mostrar,"[ENVIADO] --> id: %d\t%s\n",msgLider.idBroker, imprimirCodigo(msgLider.estado));
 			   write(fileno(stdout), mostrar, strlen(mostrar));
            }
        } while((msgLider.estado != LIDER) && (msgLider.estado!=FIN));
        
-       sprintf(mostrar,"[BROKER] --> ID LIDER ENCONTRADO: %lu\t%s\n",msgLider.idBroker, imprimirCodigo(msgLider.estado));
+       sprintf(mostrar,"[BROKER] --> ID LIDER ENCONTRADO: %d\t%s\n",msgLider.idBroker, imprimirCodigo(msgLider.estado));
        set_lider(msgLider.idBroker == *id);
        if (msgLider.idBroker == *id) printf("SOY LIDER\n");
 	   write(fileno(stdout), mostrar, strlen(mostrar));
