+++ trunk/Ejercicio 6/v5 dist. varios brokers con anillo y lider/Anillo/sender.c
@@ -17,10 +17,10 @@ main(int argc, char *argv[])
     char mostrar[300];
     
     key_t key = ftok("/tmp/buchwaldipcs", SHM_VERSION);
-    int shmversion = shmget(key, sizeof(uint64_t) , 0660 | IPC_CREAT);
-    uint64_t* id = (uint64_t*)shmat(shmversion, NULL, 0);   
+    int shmversion = shmget(key, sizeof(int) , 0660 | IPC_CREAT);
+    int* id = (int*)shmat(shmversion, NULL, 0);   
     
-    sprintf(mostrar,"[MASTER] --> ID: %lu\t| pid: %d\n",*id, getpid());
+    sprintf(mostrar,"[MASTER] --> ID: %d\t| pid: %d\n",*id, getpid());
     write(fileno(stdout),mostrar,strlen(mostrar));
     
     struct sockaddr_in addrMulticast;
@@ -220,7 +220,7 @@ main(int argc, char *argv[])
            }else{
                //sprintf(mostrar,"Recibido --> ID: %d Estado: %s\n",msgLider.idBroker,imprimirCodigo(msgLider.estado));
                //write(fileno(stdout),mostrar,strlen(mostrar));
-               sprintf(mostrar, "[RECIBIDO] <-- ID: %lu\t%s\n",msgLider.idBroker,imprimirCodigo(msgLider.estado));
+               sprintf(mostrar, "[RECIBIDO] <-- ID: %d\t%s\n",msgLider.idBroker,imprimirCodigo(msgLider.estado));
 	       write(fileno(stdout), mostrar, strlen(mostrar));
 	       //Si el id que le llego es mayor que el del actual entonces se lo envia a su siguiente
                if((msgLider.idBroker > *id) && (msgLider.estado!=FIN)){
@@ -259,7 +259,7 @@ main(int argc, char *argv[])
            }
        }while((msgLider.estado != LIDER) && (msgLider.estado!=FIN));
        
-       sprintf(mostrar, "[MASTER] --> ID LIDER ENCONTRADO: %lu\t%s\n", msgLider.idBroker, imprimirCodigo(msgLider.estado));
+       sprintf(mostrar, "[MASTER] --> ID LIDER ENCONTRADO: %d\t%s\n", msgLider.idBroker, imprimirCodigo(msgLider.estado));
        set_lider(msgLider.idBroker == *id);
        write(fileno(stdout), mostrar, strlen(mostrar));
